chore: some style updates to attendance tables
